LayoutTests/http/tests/cookies/cookie-store-get-secure.https.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/cookies-test-pre.js'></script>
<script src='../../../resources/js-test-pre.js'></script>
</head>
<body>
<script>
description('Tests that if the Cookie Store API get()/getAll() functions are used to fetch an unsecure cookie, they will set it to secure');
jsTestIsAsync = true;

onload = async () => {
    internals.setCookie({ name: "cookieName", value: "cookieValue", domain: "127.0.0.1", path: "/", isSameSiteStrict: true });

    let cookies = internals.getCookies();
    let foundCookie = false;
    for (let cookie of cookies) {
        if (foundCookie)
            break;

        if (cookie.name !== "cookieName")
            continue;

        foundCookie = true;
        if (cookie.isSecure) {
            testFailed("Set cookie should not be secure");
            finishJSTest();
        }
    }

    if (!foundCookie) {
        testFailed("Cookie was not set");
        finishJSTest();
    }

    let cookie = await cookieStore.get("cookieName");
    if (cookie.secure)
        testPassed("Cookie fetched by CookieStore::get() is secure");
    else
        testFailed("Cookie fetched by CookieStore::get() is not secure");

    cookies = await cookieStore.getAll("cookieName");
    if (cookies[0].secure)
        testPassed("Cookie fetched by CookieStore::getAll() is secure");
    else
        testFailed("Cookie fetched by CookieStore::getAll() is not secure");

    finishJSTest();
};
</script>
<script src='resources/cookies-test-post.js'></script>
</body>
</html>
