LayoutTests/http/tests/cookies/cookie-store-get-secure.https-expected.txt

Tests that if the Cookie Store API get()/getAll() functions are used to fetch an unsecure cookie, they will set it to secure

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Cookie fetched by CookieStore::get() is secure
PASS Cookie fetched by CookieStore::getAll() is secure
PASS successfullyParsed is true

TEST COMPLETE
